script: Use with_layout_state during querySelector calls
